Move NFTokenID type out of models
